support mix items not in store
